rename SanityEØSBegrunnelse.kt -> RestSanityEØSBegrunnelse.kt Flyttet RestSanityEØSBegrunnelse til brev/domene/eos La til ovrigeTriggere i RestSanityEØSBegrunnelse
