+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/domene/eøs/EØSBegrunnelseUtil.kt
@@ -3,7 +3,6 @@ package no.nav.familie.ba.sak.kjerne.brev.domene.eøs
 import no.nav.familie.ba.sak.kjerne.brev.domene.MinimertKompetanse
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.KompetanseAktivitet
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.KompetanseResultat
-import no.nav.familie.ba.sak.kjerne.vedtak.begrunnelser.BarnetsBostedsland
 
 fun hentKompetanserForEØSBegrunnelse(
     eøsBegrunnelseMedTriggere: EØSBegrunnelseMedTriggere,
